requirements.txt
bitarray==2.8.1
bitstring==4.1.1
docopt==0.6.2
h26x-extractor==0.8.1
pyusb==1.2.1
regex==2023.8.8
tabulate==0.9.0
